+++ src/components/about/carousel-project-participants.svelte
@@ -1,47 +1,6 @@
 <script lang="ts">
-	import { AccordionItem, tableMapperValues } from '@skeletonlabs/skeleton';
-	import { Table } from '@skeletonlabs/skeleton';
-	import type { TableSource } from '@skeletonlabs/skeleton';
-	import type { PersonData } from '$domains';
-	import { personsData } from '../../static/components/index.ts';
-	import { ListProjectParticipants } from '$components';
-	import { ArrowLeftIcon, ArrowRightIcon } from '$images';
-
-	// Helper function to create TableSource objects
-	function createTableSource(data: PersonData[]): TableSource {
-		return {
-			head: ['Bezeichnung', 'Beschreibung'],
-			body: tableMapperValues(data, ['description', 'value']),
-			meta: tableMapperValues(data, ['position', 'description', 'value'])
-		};
-	}
-	// Creating TableSource objects for each person in a map
-	const tableSources = personsData.map((person) => ({
-		...person, // Person takes the entire map from personsData
-		table: createTableSource(person.data) // Adds an element through the function
-	}));
-	// Carousel element
-	let elemCarousel: HTMLDivElement;
-
-	// Carousel navigation functions
-	function carouselMove(direction: 'left' | 'right'): void {
-		const step = elemCarousel.clientWidth;
-		const maxScroll = elemCarousel.scrollWidth - step;
-		let newX =
-			direction === 'left'
-				? Math.max(0, elemCarousel.scrollLeft - step)
-				: Math.min(maxScroll, elemCarousel.scrollLeft + step);
-
-		elemCarousel.scroll(newX, 0);
-	}
-
-	function carouselLeft(): void {
-		carouselMove('left');
-	}
-
-	function carouselRight(): void {
-		carouselMove('right');
-	}
+	import { CarouselElement } from '$components';
+	import { AccordionItem } from '@skeletonlabs/skeleton';
 </script>
 
 <AccordionItem open>
@@ -61,47 +20,7 @@
 		</h2></svelte:fragment
 	>
 	<svelte:fragment slot="content">
-		<!-- Carousel -->
-		<div class="card p-4 grid grid-cols-[auto_1fr_auto] gap-4 items-center">
-			<!-- Button: Left -->
-			<button
-				type="button"
-				class="btn-icon variant-filled-secondary dark:variant-filled-primary"
-				on:click={carouselLeft}
-			>
-				<ArrowLeftIcon />
-			</button>
-			<!-- Card consisting of table and image -->
-			<div
-				bind:this={elemCarousel}
-				class="snap-x snap-mandatory scroll-smooth flex overflow-x-auto"
-			>
-				{#each tableSources as source}
-					<div class="snap-center rounded-container-token carousel-item">
-						<!-- Image part of the card -->
-						<div class="image-container items-center">
-							<img class="carousel-image" src={source.src} alt="Bild von Person" loading="lazy" />
-						</div>
-						<!-- Text part of the card - Table for Desktop and List for mobile-->
-						<div class="table-container tableData">
-							<Table source={source.table} />
-						</div>
-						<div class="table-container flex tableDataText">
-							<ListProjectParticipants source={source.data} />
-						</div>
-					</div>
-					<span class="divider-vertical h-100" />
-				{/each}
-			</div>
-			<!-- Button: Right -->
-			<button
-				type="button"
-				class="btn-icon variant-filled-secondary dark:variant-filled-primary"
-				on:click={carouselRight}
-			>
-				<ArrowRightIcon />
-			</button>
-		</div>
+		<CarouselElement />
 	</svelte:fragment>
 </AccordionItem>
 
@@ -119,59 +38,10 @@
 		--logo-icon-size: 2.188rem; /* Size for logo icons */
 	}
 
-	/* Styles for carousel items */
-	.carousel-item {
-		display: flex;
-		flex-direction: row;
-		width: 100%;
-		justify-content: center;
-		min-width: var(--carousel-item-min-width);
-		margin-right: var(--carousel-item-margin);
-		margin-left: var(--carousel-item-margin);
-	}
-
-	/* Styles for table containers */
-	.table-container {
-		text-align: left;
-		width: 100%;
-		word-wrap: break-word;
-		min-width: var(--table-container-min-width);
-		min-height: var(--table-container-min-height);
-	}
-
-	/* Styles for image containers */
-	.image-container {
-		width: 100%;
-		display: flex;
-		justify-content: center;
-		min-width: var(--image-container-min-width);
-		min-height: var(--image-container-min-height);
-		margin-right: 20px;
-	}
-
-	/* Styles for images in the carousel */
-	.carousel-image {
-		border-radius: 50%;
-		overflow: hidden;
-		width: 100%;
-		height: auto;
-		object-fit: cover; /* Ensures the image fills the container */
-		max-width: var(--carousel-image-max-width);
-		max-height: var(--carousel-image-max-height);
-	}
-
 	/* Styles for various icons */
 	.icon-team {
 		height: var(--logo-icon-size);
 		width: var(--logo-icon-size);
-	}
-
-	.tableData {
-		display: flex;
-	}
-
-	.tableDataText {
-		display: none;
 	}
 
 	/* Media Query for small screens (e.g., mobile phones) */
@@ -194,40 +64,8 @@
 			overflow-wrap: break-word; /* Alternative property for line break */
 		}
 
-		.carousel-image {
-			width: 9.375rem; /* Reduced size for images and icons */
-			height: 9.375rem;
-			margin-bottom: 10px;
-			margin-left: 0px;
-		}
-
 		.icon-team {
 			display: none;
-		}
-
-		.carousel-item {
-			flex-direction: column; /* Changes the alignment of flex items */
-		}
-
-		/* Adjustments for text and buttons */
-		button {
-			font-size: 0.9rem; /* Smaller font size for normal text and buttons */
-		}
-
-		.btn-icon {
-			display: none;
-		}
-
-		.table-container {
-			text-align: center;
-		}
-
-		.tableData {
-			display: none;
-		}
-
-		.tableDataText {
-			display: flex;
 		}
 	}
 </style>
