+++ src/components/index.ts
@@ -9,6 +9,7 @@ export { default as LegalNotice } from './about/legal-notice.svelte';
 export { default as ProductVideo } from './about/product-video.svelte';
 export { default as ProjectDocumentation } from './about/project-documentation.svelte';
 export { default as CarouselProjectParticipants } from './about/carousel-project-participants.svelte';
+export { default as CarouselElement } from './about/carousel.svelte';
 export { default as MediaButtonsLogo } from './about/media-buttons-logo.svelte';
 export { default as ListProjectParticipants } from './about/list-project-participants.svelte';
